Moved dialog tests into a dialogs folder.
